third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/frame-src-child-frame-navigates-to-blocked-origin.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="frame-src https://localhost:8443">
    <script>
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
    </script>
</head>
<body>
    The test verifies that Content-Security-Policy from the main frame
    restricts child frame's location even when the location is changed
    as a result of a navigation trigerred from within the child frame
    (which might reside in another renderer process due to --site-per-process).
    <iframe src="https://localhost:8443/security/contentSecurityPolicy/resources/frame-that-navigates-itself.html"></iframe>
</body>
</html>
